recipe-app-api/app/requirements.txt
Django
djangorestframework
psycopg2-binary
flake8 
